Refactor postion loading to separate file
